Combined models into single file models.py and moved to db directory. Removed unnecessary files. Closes #10
